refactor: expose the runtime error boundary for verification
